requirements.txt
Pillow
s3transfer
